refactor(S44): H4 — extract EduSectionTitle + EduDisclaimer to shared widget
Moves duplicated _SectionTitle and _Disclaimer from
OptimisationDecaissementScreen and SuccessionPatrimoineScreen
into widgets/coach/edu_shared_widgets.dart (EduSectionTitle, EduDisclaimer).
