+++ apps/mobile/lib/widgets/coach/edu_shared_widgets.dart
@@ -1,0 +1,61 @@
+/// Shared widgets for educational coach screens (65+).
+///
+/// Used by: OptimisationDecaissementScreen, SuccessionPatrimoineScreen.
+library;
+
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:mint_mobile/theme/colors.dart';
+
+/// Section title — Montserrat 15 bold.
+class EduSectionTitle extends StatelessWidget {
+  final String text;
+  const EduSectionTitle({super.key, required this.text});
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      text,
+      style: GoogleFonts.montserrat(
+        fontSize: 15,
+        fontWeight: FontWeight.w700,
+        color: MintColors.textPrimary,
+      ),
+    );
+  }
+}
+
+/// LSFin disclaimer banner (yellow background, info icon).
+class EduDisclaimer extends StatelessWidget {
+  final String text;
+  const EduDisclaimer({super.key, required this.text});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.all(14),
+      decoration: BoxDecoration(
+        color: const Color(0xFFFFF8E1),
+        borderRadius: BorderRadius.circular(12),
+        border: Border.all(color: const Color(0xFFFFE082)),
+      ),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const Icon(Icons.info_outline, size: 16, color: Color(0xFFF57F17)),
+          const SizedBox(width: 8),
+          Expanded(
+            child: Text(
+              text,
+              style: GoogleFonts.inter(
+                fontSize: 11,
+                color: const Color(0xFF5D4037),
+                height: 1.5,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
